# packages/client/spec/openapi.yaml
...
openapi: 3.1.0
info:
  title: Ownware Gateway — core wire contract
  version: 0.29.0
  description: >
    Drive a Ownware agent from any language: start a run, stream its
    events, answer permission prompts. Bearer auth applies to every
    route except /health whenever the gateway has auth enabled (always
    true on non-loopback binds; token at <dataDir>/gateway-token).
servers:
  - url: http://localhost:3011
    description: Default local gateway (ownware serve prints yours)
security:
  - bearer: []
paths:
...
  /api/v1/sources:
    post:
      operationId: registerSource
      summary: Register one logical source using safe control metadata only.
      description: >
        Requires a delegated principal; workspace and profile scope come only
        from the verified bearer. The request cannot carry paths, URLs, bytes,
        storage locators, or workspace/profile authority. Equal metadata under
        distinct idempotency keys remains distinct logical sources.
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SourceRegistrationInput' }
      responses:
        '202':
          description: Pending safe manifest. No content or storage detail is returned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceManifest' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/SourceCreationConflictResponse' }
        '413': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }
    get:
      operationId: listSources
      summary: List one bounded page of safe manifests in the verified scope.
      description: Requires a delegated principal. An empty scope returns an empty page.
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - in: query
          name: cursor
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Bounded source page without raw content or storage detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceList' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/sources/{sourceId}:
    get:
      operationId: getSource
      summary: Read one safe source manifest in the verified scope.
      description: Cross-workspace/profile identities are indistinguishable from absence.
      parameters:
        - in: path
          name: sourceId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Safe source control manifest.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceManifest' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/source-data-views/{dataViewId}:
    get:
      operationId: getSourceDataView
      summary: Read one scoped content-free Data View manifest.
      description: >
        Requires the separate source_data_views.read authority. Cross-workspace/profile
        identities and views whose source is deleting are indistinguishable from absence.
        Field labels are untrusted source header labels; no cells, row values, private
        placement, diagnostics, grant, or query authority are returned.
      parameters:
        - { in: path, name: dataViewId, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Validated Data View lineage, policy, field shape, counts, and freshness only.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content: { application/json: { schema: { $ref: '#/components/schemas/SourceDataViewManifest' } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/source-data-views/{dataViewId}/query:
    post:
      operationId: querySourceDataView
      summary: Read one exact field list and bounded row window after live grant evaluation.
      description: >
        Requires a subject-bound delegated principal carrying source_data_views.query.
        Workspace, profile, subject, purpose and channel come only from verified signed
        claims; the body cannot override identity or authority. The live grant, strict
        current lineage/lifecycle floor and artifact integrity are checked before and after the
        bounded selection. Protected denial, stale/deleting state, tamper and race failures
        return no cells through one generic unavailable response. This is a retry-safe read
        POST with no durable side effect and no SQL, filter, sort, aggregation, expression
        or arbitrary row predicate.
      parameters:
        - { in: path, name: dataViewId, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuerySourceDataViewInput' }
      responses:
        '200':
          description: Verified current cells with complete immutable lineage and bounded-window truth.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content: { application/json: { schema: { $ref: '#/components/schemas/ProtectedSourceDataViewSelection' } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '413': { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/source-resources/{resourceId}/content:
    post:
      operationId: readSourceContent
      summary: Read one bounded UTF-8 range after live grant and source-state evaluation.
      description: >
        Requires a delegated principal with source_content.read. Workspace, profile,
        subject, purpose and channel come only from that principal. Owner tokens cannot bypass
        the route or grant. Grant, lineage, availability, freshness and deletion truth
        are checked before and after the private byte read; denial is indistinguishable
        from absence.
      parameters:
        - { in: path, name: resourceId, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReadSourceContentInput' }
      responses:
        '200':
          description: Verified current source bytes and safe immutable lineage.
          content: { application/json: { schema: { $ref: '#/components/schemas/ProtectedSourceContent' } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '413': { $ref: '#/components/responses/ErrorResponse' }
...
components:
  responses:
    ErrorResponse:
      description: Safe typed error. Raw internal/provider detail is not returned.
      headers:
        X-Ownware-Correlation-Id:
          schema: { type: string, format: uuid }
          description: Opaque server-generated support correlation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    SourceCreationConflictResponse:
      description: >
        Safe typed creation conflict. `source_quota_exceeded` reports only the
        blocked resource class plus standard safe error metadata; it never
        reports usage, failed workspace/profile scope, resource identities,
        paths, or a retry delay.
      headers:
        X-Ownware-Correlation-Id:
          schema: { type: string, format: uuid }
          description: Opaque server-generated support correlation.
      content:
        application/json:
          schema:
            oneOf:
              - allOf:
                  - { $ref: '#/components/schemas/ErrorResponse' }
                  - not:
                      required: [error]
                      properties:
                        error: { const: source_quota_exceeded }
              - { $ref: '#/components/schemas/SourceQuotaExceededError' }
...
  schemas:
...
    SourceRegistrationInput:
      type: object
      required: [kind, label, classification, authority, audiencePolicyRef, sensitivityPolicyRef, purposePolicyRef, retentionPolicyRef, freshnessPolicyRef]
      properties:
        kind:
          type: string
          enum: [file, text, visual, structured_export, cloud_document, connected_snapshot, supported_other]
        label:
          type: string
          minLength: 1
          maxLength: 160
          description: Safe single-line display label; control characters reject.
        classification: { type: string, enum: [public, internal, confidential, restricted] }
        authority: { type: string, enum: [source_of_record, supporting_reference, example, excluded] }
        audiencePolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        sensitivityPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        purposePolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        retentionPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        freshnessPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
      additionalProperties: false
    SourceHealth:
      type: object
      required: [registration, inspection, preparation, access, freshness, conflict, deletion]
      properties:
        registration: { type: string, enum: [pending, registered, rejected] }
        inspection: { type: string, enum: [not_started, queued, inspecting, complete, partial, failed] }
        preparation: { type: string, enum: [not_requested, queued, preparing, ready, partial, failed] }
        access: { type: string, enum: [available, denied, expired, disconnected, wrong_identity] }
        freshness: { type: string, enum: [fresh, aging, stale, unknown] }
        conflict: { type: string, enum: [none, suspected, confirmed, resolved] }
        deletion: { type: string, enum: [active, frozen, deleting, partially_deleted, deleted] }
      additionalProperties: false
    SourceManifest:
      type: object
      required: [sourceId, kind, label, classification, authority, audiencePolicyRef, sensitivityPolicyRef, purposePolicyRef, retentionPolicyRef, freshnessPolicyRef, revision, currentVersionId, health, createdAt, updatedAt]
      properties:
        sourceId: { type: string, format: uuid }
        kind:
          type: string
          enum: [file, text, visual, structured_export, cloud_document, connected_snapshot, supported_other]
        label: { type: string, minLength: 1, maxLength: 160 }
        classification: { type: string, enum: [public, internal, confidential, restricted] }
        authority: { type: string, enum: [source_of_record, supporting_reference, example, excluded] }
        audiencePolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        sensitivityPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        purposePolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        retentionPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        freshnessPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        revision: { type: integer, minimum: 1 }
        currentVersionId: { type: [string, 'null'], format: uuid }
        health: { $ref: '#/components/schemas/SourceHealth' }
        createdAt: { type: integer, minimum: 0 }
        updatedAt: { type: integer, minimum: 0 }
      additionalProperties: false
    SourceList:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/SourceManifest' }
        nextCursor: { type: [string, 'null'], format: uuid }
      additionalProperties: false
...
    SourceDataViewField:
      type: object
      required: [fieldId, ordinal, label]
      properties:
        fieldId: { type: string, pattern: '^field\.[0-9a-f]{32}$' }
        ordinal: { type: integer, minimum: 0, maximum: 255 }
        label:
          type: string
          minLength: 1
          maxLength: 65536
          description: Untrusted display label copied mechanically from the source header.
      additionalProperties: false
    SourceDataViewManifest:
      type: object
      required: [dataViewId, jobId, sourceId, sourceVersionId, implementationVersion, sourceRevision, sourceChecksum, artifactChecksum, artifactByteCount, fieldCount, rowCount, fields, classification, authority, audiencePolicyRef, sensitivityPolicyRef, purposePolicyRef, retentionPolicyRef, freshnessPolicyRef, freshness, createdAt, staleAt]
      properties:
        dataViewId: { type: string, format: uuid }
        jobId: { type: string, format: uuid }
        sourceId: { type: string, format: uuid }
        sourceVersionId: { type: string, format: uuid }
        implementationVersion: { type: string, const: csv_data_view.v1 }
        sourceRevision: { type: integer, minimum: 1 }
        sourceChecksum: { type: string, pattern: '^sha256:[0-9a-f]{64}$' }
        artifactChecksum: { type: string, pattern: '^sha256:[0-9a-f]{64}$' }
        artifactByteCount: { type: integer, minimum: 1, maximum: 134217728 }
        fieldCount: { type: integer, minimum: 1, maximum: 256 }
        rowCount: { type: integer, minimum: 0, maximum: 100000 }
        fields:
          type: array
          minItems: 1
          maxItems: 256
          items: { $ref: '#/components/schemas/SourceDataViewField' }
        classification: { type: string, enum: [public, internal, confidential, restricted] }
        authority: { type: string, enum: [source_of_record, supporting_reference, example] }
        audiencePolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        sensitivityPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        purposePolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        retentionPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        freshnessPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        freshness: { type: string, enum: [current, stale] }
        createdAt: { type: integer, minimum: 0 }
        staleAt: { type: [integer, 'null'], minimum: 0 }
      additionalProperties: false
...
    QuerySourceDataViewInput:
      type: object
      required: [consent, fieldIds, rowOffset, rowCount]
      properties:
        consent: { $ref: '#/components/schemas/AccessConsent' }
        fieldIds:
          type: array
          minItems: 1
          maxItems: 32
          uniqueItems: true
          items: { type: string, pattern: '^field\.[0-9a-f]{32}$' }
        rowOffset: { type: integer, minimum: 0, maximum: 99999 }
        rowCount: { type: integer, minimum: 1, maximum: 256 }
      additionalProperties: false
      description: >
        The product of fieldIds and rowCount cannot exceed 8192. Identity,
        purpose, channel, operation, autonomy and permission mode are never body fields.
    SourceDataViewSelectionRow:
      type: object
      required: [rowId, ordinal, values]
      properties:
        rowId: { type: string, pattern: '^row\.[0-9a-f]{32}$' }
        ordinal: { type: integer, minimum: 0, maximum: 99999 }
        values:
          type: array
          minItems: 1
          maxItems: 32
          items:
            type: string
            description: Inert untrusted source data; never interpreted as an instruction or expression.
      additionalProperties: false
    ProtectedSourceDataViewSelection:
      type: object
      required: [dataViewId, sourceId, sourceVersionId, sourceRevision, sourceChecksum, artifactChecksum, freshness, classification, authority, implementationVersion, rowOffset, requestedRowCount, returnedRowCount, totalRowCount, complete, fields, rows, observedAt]
      properties:
        dataViewId: { type: string, format: uuid }
        sourceId: { type: string, format: uuid }
        sourceVersionId: { type: string, format: uuid }
        sourceRevision: { type: integer, minimum: 1 }
        sourceChecksum: { type: string, pattern: '^sha256:[0-9a-f]{64}$' }
        artifactChecksum: { type: string, pattern: '^sha256:[0-9a-f]{64}$' }
        freshness: { type: string, const: current }
        classification: { type: string, enum: [public, internal, confidential, restricted] }
        authority: { type: string, enum: [source_of_record, supporting_reference, example] }
        implementationVersion: { type: string, const: csv_data_view_selection.v1 }
        rowOffset: { type: integer, minimum: 0, maximum: 99999 }
        requestedRowCount: { type: integer, minimum: 1, maximum: 256 }
        returnedRowCount: { type: integer, minimum: 1, maximum: 256 }
        totalRowCount: { type: integer, minimum: 1, maximum: 100000 }
        complete:
          type: boolean
          description: True only when the returned window reaches the current final row.
        fields:
          type: array
          minItems: 1
          maxItems: 32
          items: { $ref: '#/components/schemas/SourceDataViewField' }
        rows:
          type: array
          minItems: 1
          maxItems: 256
          items: { $ref: '#/components/schemas/SourceDataViewSelectionRow' }
        observedAt: { type: integer, minimum: 0 }
      additionalProperties: false
    AccessConsent:
      oneOf:
        - type: object
          required: [state]
          properties:
            state: { type: string, const: not_required }
          additionalProperties: false
        - type: object
          required: [state, evidenceId]
          properties:
            state: { type: string, const: recorded }
            evidenceId:
              type: string
              pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
          additionalProperties: false
...
    ReadSourceContentInput:
      type: object
      required: [consent, byteStart, byteEnd]
      properties:
        consent: { $ref: '#/components/schemas/AccessConsent' }
        byteStart: { type: integer, minimum: 0 }
        byteEnd: { type: integer, minimum: 1 }
      additionalProperties: false
      description: Subject identity comes only from the verified delegated principal.
    ProtectedSourceContent:
      type: object
      required: [resourceId, sourceId, sourceVersionId, sourceRevision, sourceChecksum, resourceChecksum, freshness, classification, authority, text, byteStart, byteEnd, byteCount, totalByteCount, observedAt]
      properties:
        resourceId: { type: string, format: uuid }
        sourceId: { type: string, format: uuid }
        sourceVersionId: { type: string, format: uuid }
        sourceRevision: { type: integer, minimum: 1 }
        sourceChecksum: { type: string, pattern: '^sha256:[0-9a-f]{64}$' }
        resourceChecksum: { type: string, pattern: '^sha256:[0-9a-f]{64}$' }
        freshness: { type: string, const: current }
        classification: { type: string, enum: [public, internal, confidential, restricted] }
        authority: { type: string, enum: [source_of_record, supporting_reference, example] }
        text: { type: string }
        byteStart: { type: integer, minimum: 0 }
        byteEnd: { type: integer, minimum: 1 }
        byteCount: { type: integer, minimum: 1, maximum: 65536 }
        totalByteCount: { type: integer, minimum: 1 }
        observedAt: { type: integer, minimum: 0 }
      additionalProperties: false
...
    ErrorResponse:
      type: object
      required: [error, message, category, correlationId]
      properties:
        error: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        message: { type: string, maxLength: 500 }
        category: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        correlationId: { type: string, format: uuid }
        retryAfter:
          type: number
          minimum: 0
          description: Optional retry delay in seconds.
        actualRevision:
          type: integer
          minimum: 1
          description: Safe actual source revision, present only for `source_upload_refresh_conflict`.
        actualCurrentVersionId:
          type: [string, 'null']
          format: uuid
          description: Safe actual current source version, present only for `source_upload_refresh_conflict`.
...
    SourceQuotaExceededError:
      type: object
      required: [error, message, category, correlationId, resourceClass]
      description: >
        Quota denial without usage, failed scope, source/upload/job/resource
        identities, paths, or retry timing.
      properties:
        error: { type: string, const: source_quota_exceeded }
        message: { type: string, const: Source quota does not allow this operation. }
        category: { type: string, const: invalid_request }
        correlationId: { type: string, format: uuid }
        resourceClass:
          type: string
          enum: [source_registrations, source_storage_bytes, source_upload_sessions, source_jobs, source_derived_resources]
      additionalProperties: false
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: >
        Gateway session token. Enabled automatically on any non-loopback
        bind; persisted at <dataDir>/gateway-token (0600).